+++ src/Foundation/Support/helpers.php
@@ -3,6 +3,7 @@
 declare(strict_types=1);
 
 use Vanilo\MasterProduct\Contracts\MasterProduct;
+use Vanilo\MasterProduct\Contracts\MasterProductVariant;
 
 /**
  * Returns the price formatted
@@ -23,4 +24,9 @@ function format_price($price, string $currency = null)
 function is_master_product(object $product): bool
 {
     return $product instanceof MasterProduct;
+}
+
+function is_master_product_variant(object $product): bool
+{
+    return $product instanceof MasterProductVariant;
 }